% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{table1 function}
\usage{
tab1(
  DESIGN,
  STRATA = NULL,
  OVERALL = FALSE,
  CONDIGITS = 1,
  CATDIGITS = 1,
  FUNC = "mean",
  TYPE = 1,
  MARGIN = 2,
  MODE = 1
)
}
\arguments{
\item{DESIGN:}{put survey data which is svydesign() from survey package}

\item{STRATA:}{put stratification varaible}

\item{OVERALL:}{if you want overall statistics, set TRUE}

\item{CONDIGITS:}{digits of continuous variable}

\item{CATDIGITS:}{digits of categorical variable}

\item{FUNC:}{"mean" or "median"}

\item{TYPE:}{type of print FUNC}

\item{MARGIN:}{margin of percentage. 1 as row percent, and 2 as col percent}

\item{MODE:}{mode of descriptive statistics. 1 as consider weights of complex design, 2 as do not consider weights for n size, 3 as do not consider complex design.}
}
\description{
This function create table1 in biostatistics
}
\examples{
options(survey.lonely.psu = "adjust")
options(survey.adjust.domain.lonely = TRUE)
ddd <- select(apiclus2, dnum,pw,fpc1,api00,api99,target,sch.wide,comp.imp,meals)
ddd <- preprocess(ddd)
ddd$fpc1 <- as.numeric(as.character(ddd$fpc1))
dclus1 <- svydesign(id=~dnum, weights=~pw, data=ddd, fpc=~fpc1)
tab1(dclus1, MODE = 1)
tab1(dclus1, MODE = 2)
tab1(dclus1, MODE = 3)
tab1(dclus1, STRATA = "sch.wide")
tab1(dclus1, STRATA = "sch.wide", MODE = 1)
tab1(dclus1, STRATA = "sch.wide", MODE = 2)
tab1(dclus1, STRATA = "sch.wide", MODE = 3)
tab1(dclus1, STRATA = "sch.wide", MARGIN = 1, OVERALL = T)
}
\keyword{complex}
\keyword{create}
\keyword{design}
\keyword{in}
\keyword{survey}
\keyword{table1}
